QLogging: fix build: narrowing conversion from qsizetype to size_t
Amends a6070847f075296458e1afef0210f9c89aca4b2a.

error: narrowing conversion of ‘(& ba)->QByteArray::size()’ from ‘qsizetype’ {aka ‘long long int’} to ‘size_t’ {aka ‘long unsigned int’} [-Werror=narrowing]

Task-number: QTBUG-120047
Task-number: QTBUG-120048

+++ src/corelib/global/qlogging.cpp
@@ -1860,7 +1860,7 @@ static bool systemd_default_message_handler(QtMsgType type,
                                  ? "QT_CATEGORY="_ba + context.category : QByteArray();
 
     auto toIovec = [](const QByteArray &ba) {
-        return iovec{const_cast<char*>(ba.data()), ba.size()};
+        return iovec{ const_cast<char*>(ba.data()), size_t(ba.size()) };
     };
 
     struct iovec fields[7] = {
